improve folder structure with import path update
