+++ web-app/src/components/agents/agent-modal/card-sections/card-section.tsx
@@ -1,0 +1,9 @@
+function CardSection({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="bg-muted/80 border-foreground/20 flex w-full flex-col gap-6 rounded-xl border-1 p-6 backdrop-blur-3xl">
+      {children}
+    </div>
+  );
+}
+
+export { CardSection };
